# Fix sampler nan check when calling top_k_top_p_sampling_from_probs