--- a/python/sglang/srt/layers/sampler.py
+++ b/python/sglang/srt/layers/sampler.py
@@ -112,10 +112,6 @@ class Sampler(nn.Module):
                         check_nan=self.use_nan_detection,
                     )
 
-                    # if self.use_nan_detection and not torch.all(success):
-                    #     logger.warning("Detected errors during sampling!")
-                    #     batch_next_token_ids = torch.zeros_like(batch_next_token_ids)
-
             elif global_server_args_dict["sampling_backend"] == "pytorch":
                 # A slower fallback implementation with torch native operations.
                 batch_next_token_ids = top_k_top_p_min_p_sampling_from_probs_torch(