<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

~~Remove success tensor that is not returned from top_k_top_p_sampling_from_probs of the latest flash infer api anymore.~~

The current code on main branch will throw variable referenced before assigned error if `self.use_nan_detection` evaluate to true. This PR fixes it.

We don't need to check for success anymore as the PR in flash infer has improved the robustness of the sampling. Now tokens are guaranteed to be generated.
Details in: https://github.com/flashinfer-ai/flashinfer/pull/912

Please refer to flash infer for reference:
https://github.com/flashinfer-ai/flashinfer/blob/main/flashinfer/sampling.py


## Modifications

Remove successor, pass in check for nan flag, and fix the function signature.

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.


## Test Result
No latency regression after the change.

```python3 -m unittest test_bench_one_batch.TestBenchOneBatch.test_bs1```

Before change:
```
Warmup ...
Prefill. latency: 0.33793 s, throughput:    378.77 token/s
Decode.  median latency: 0.00587 s, median throughput:    170.22 token/s
Total. latency:  0.764 s, throughput:    178.08 token/s
Benchmark ...
Prefill. latency: 0.01321 s, throughput:   9686.79 token/s
Decode.  median latency: 0.00580 s, median throughput:    172.50 token/s
Total. latency:  0.054 s, throughput:   2526.44 token/s
```
After change:
```
Warmup ...
Prefill. latency: 0.50365 s, throughput:    254.14 token/s
Decode.  median latency: 0.00589 s, median throughput:    169.79 token/s
Total. latency:  1.031 s, throughput:    131.89 token/s
Benchmark ...
Prefill. latency: 0.01323 s, throughput:   9676.49 token/s
Decode.  median latency: 0.00579 s, median throughput:    172.60 token/s
Total. latency:  0.054 s, throughput:   2527.25 token/s
```