@yinfan98 2025-04-19
U can fix sgl-kernel/python/sgl_kernel/sampling.py firstly, for python/sglang/srt/layers/sampler.py we need bump sgl-kernel to new version.