@yubofredwang 2025-04-19
> U can fix sgl-kernel/python/sgl_kernel/sampling.py firstly, for python/sglang/srt/layers/sampler.py we need bump sgl-kernel to new version.

Yes, but the change in sgl_kernel is not required for the code to work. python does not do type checking. so returning `torch.Tensor` or a Tuple won't affect code execution.

This change is just to keep things consistent with flash infer